requirements.txt
git+https://github.com/django/django.git@stable/1.6.x
psycopg2
django-categories==1.2.2
django-registration==1.0
django-imagekit==3.0.3
django-softdelete==0.3.4
django-hstore==1.1.1
